Externalized Card component
