+++ geonode_mapstore_client/client/js/components/Card/index.js
@@ -1,0 +1,2 @@
+import Cards from './Card';
+export default Cards;
